+++ net-gateway/src/endpoints/charts/http_responses_distribution/response/http_responses_distribution_bucket.rs
@@ -1,10 +1,10 @@
-use net_reporter_api::api::http_responses_dist::http_responses_bucket;
+use net_reporter_api::api::http_responses_distribution::http_responses_distribution_bucket;
 
 use serde::Deserialize;
 use serde::Serialize;
 
 #[derive(Serialize, Deserialize, Debug, Clone, Default)]
-pub struct HttpResponsesDistBucketResponse {
+pub struct HttpResponsesDistributionBucketResponse {
     #[serde(rename = "bucketTimestamp")]
     bucket_timestamp: u64,
     #[serde(rename = "responseCode")]
@@ -12,7 +12,7 @@ pub struct HttpResponsesDistBucketResponse {
     amount: u64,
 }
 
-impl HttpResponsesDistBucketResponse {
+impl HttpResponsesDistributionBucketResponse {
     pub fn new(bucket_timestamp: u64, response_code: u64, amount: u64) -> Self {
         Self {
             bucket_timestamp,
@@ -22,8 +22,8 @@ impl HttpResponsesDistBucketResponse {
     }
 }
 
-impl From<http_responses_bucket::HttpResponsesBucketDTO> for HttpResponsesDistBucketResponse {
-    fn from(bucket: http_responses_bucket::HttpResponsesBucketDTO) -> Self {
+impl From<http_responses_distribution_bucket::HttpResponsesDistributionBucketDTO> for HttpResponsesDistributionBucketResponse {
+    fn from(bucket: http_responses_distribution_bucket::HttpResponsesDistributionBucketDTO) -> Self {
         Self {
             bucket_timestamp: bucket.get_bucket_timestamp() as u64,
             response_code: bucket.get_response_code() as u64,