+++ net-gateway/src/endpoints/charts/http_responses_distribution/response/mod.rs
@@ -1,0 +1,3 @@
+pub mod http_responses_distribution_bucket;
+pub mod http_responses_distribution;
+